Still not able to get ball to move might have to refactor
