+++ Breakout/src/Brick/Brick.java
@@ -1,10 +1,13 @@
-package Main;
+package Brick;
+
+import Main.Shapes;
 
 import java.awt.*;
 import java.awt.geom.Rectangle2D;
+import java.util.ArrayList;
 
 
-public class Brick extends Shapes{
+public class Brick extends Shapes {
 
     private static final int Height = 10;
     private static final int Width = 30;
@@ -18,13 +21,16 @@ public class Brick extends Shapes{
     private Brick(Rectangle2D rectangle, Color color) {
         super(rectangle, color, true);
     }
-/*
+
+
+
     public ArrayList<Brick> add(Brick other) {
+        ArrayList<Brick> bricks = new ArrayList<Brick>();
         Rectangle2D rectangleOne = super.getBounds();
         Rectangle2D rectangleTwo = other.getBounds();
         rectangleOne.add(rectangleTwo);
-        return new ArrayList<Brick>(rectangleOne, super.getColor());
+        return bricks;
     }
 
- */
+
 }
