+++ Breakout/src/Main/Breakout.java
@@ -4,7 +4,9 @@ public class Breakout {
     public static void main(String[] args) {
 
         Frames frame = new Frames();
+
         frame.setVisible(true);
+
 
 
     }
