+++ Game-Instructions.md
@@ -3,18 +3,14 @@
 Place the instructions for how to play your game in this file.  The instructions should be clear enough for any non-technical computer user to understand and follow.
 
 
-1. Clone from github repo.
 
-2.download needed dependencies such as javafx.
 
-3.once downloaded add all javafx jar files to project.
+1.Run file and game will popup.
 
-4.Run file and game will popup.
-
-5. click buttons <ctrl+s>
+2. click buttons <ctrl+s> to start game. <ctrl+p> to pause.
    
-6. slide paddle side to side to keep ball from hitting bottom.
+3. slide paddle side to side to keep ball from hitting bottom.
    
-7.Knockout all bricks, and you will win game.
+4.Knockout all bricks, and you will win game.
