+++ UPER.md
@@ -9,9 +9,10 @@
 *Develop a game that has OOP and OOD properties.
 *I am to make a plan of how to implement the four pillars of OOP into a game app to show that i understand.
 *
+
+*Research options and breakdown which project fits best.
 *
-*
-*
+*learn how to implement UI and which library would be best used.
 *
 *
 <h2>
@@ -26,8 +27,6 @@
 * come up with any test to make sure game is running correctly
 *
 *
-*
-*
 <h2>
     3. Executing the Plan
 </h2>
@@ -37,9 +36,7 @@
 
 * i watched a few tutorials and read over some online sources to get started.
   
-*
-*
-*
+* add levels with different designs for bricks.
 *
 *
 <h2>
@@ -51,7 +48,7 @@
   
 *refactor all of my logic classes to run more independently.
 *
-*
-*
+
+*Do more research before jumping in to have a cleaner finish
 *
 *
